requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Werkzeug
pycountry
requests
python-dotenv
numpy
pandas
matplotlib
